requirements.txt
discord
python-dotenv
pytz

# llm
openai
langchain